move up to its own command
