+++ cmd/init.go
@@ -35,7 +35,6 @@ import (
 
 const (
 	stignore          = ".stignore"
-	defaultManifest   = "okteto.yml"
 	secondaryManifest = "okteto.yaml"
 	createDeployment  = "new deployment"
 )
@@ -71,7 +70,7 @@ func Init() *cobra.Command {
 			}
 
 			log.Success(fmt.Sprintf("Okteto manifest (%s) created", devPath))
-			if devPath == defaultManifest {
+			if devPath == utils.DefaultDevManifest {
 				log.Information("Run 'okteto up' to activate your development container")
 			} else {
 				log.Information("Run 'okteto up -f %s' to activate your development container", devPath)
@@ -81,7 +80,7 @@ func Init() *cobra.Command {
 	}
 
 	cmd.Flags().StringVarP(&namespace, "namespace", "n", "", "namespace target for generating the okteto manifest")
-	cmd.Flags().StringVarP(&devPath, "file", "f", defaultManifest, "path to the manifest file")
+	cmd.Flags().StringVarP(&devPath, "file", "f", utils.DefaultDevManifest, "path to the manifest file")
 	cmd.Flags().BoolVarP(&overwrite, "overwrite", "o", false, "overwrite existing manifest file")
 	return cmd
 }
